backend/requirements.txt
Django>=4.3
djangorestframework>=3.15
djangorestframework-simplejwt>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
django-cors-headers>=4.0
requests>=2.31
beautifulsoup4>=4.13
